+++ dflib/src/main/java/org/dflib/print/Printer.java
@@ -5,12 +5,34 @@ import org.dflib.Series;
 
 public interface Printer {
 
-    default String toString(Series<?> s) {
+    /**
+     * @since 2.0.0
+     */
+    default String print(Series<?> s) {
         return print(new StringBuilder(), s).toString();
     }
 
-    default String toString(DataFrame df) {
+    /**
+     * @since 2.0.0
+     */
+    default String print(DataFrame df) {
         return print(new StringBuilder(), df).toString();
+    }
+
+    /**
+     * @deprecated in favor of {@link #print(Series)}
+     */
+    @Deprecated(since = "2.0.0", forRemoval = true)
+    default String toString(Series<?> s) {
+        return print(s);
+    }
+
+    /**
+     * @deprecated in favor of {@link #print(DataFrame)}
+     */
+    @Deprecated(since = "2.0.0", forRemoval = true)
+    default String toString(DataFrame df) {
+        return print(df);
     }
 
     StringBuilder print(StringBuilder out, Series<?> s);