+++ dflib/src/test/java/org/dflib/print/TabularPrinter_SeriesTest.java
@@ -8,7 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class TabularPrinter_SeriesTest {
 
     @Test
-    public void toString_Double() {
+    public void print_Double() {
         TabularPrinter p = new TabularPrinter(5, 10);
 
         assertEquals(System.lineSeparator() +
@@ -16,11 +16,11 @@ public class TabularPrinter_SeriesTest {
                 "    -1.01" + System.lineSeparator() +
                 " -10000.5" + System.lineSeparator() +
                 "3965001.2" + System.lineSeparator() +
-                "4 elements", p.toString(Series.ofDouble(1.0, -1.01, -10_000.5, 3_965_001.2)));
+                "4 elements", p.print(Series.ofDouble(1.0, -1.01, -10_000.5, 3_965_001.2)));
     }
 
     @Test
-    public void toString_Int() {
+    public void print_Int() {
         TabularPrinter p = new TabularPrinter(5, 10);
 
         assertEquals(System.lineSeparator() +
@@ -28,21 +28,21 @@ public class TabularPrinter_SeriesTest {
                 "     -1" + System.lineSeparator() +
                 " -10000" + System.lineSeparator() +
                 "3965001" + System.lineSeparator() +
-                "4 elements", p.toString(Series.ofInt(1, -1, -10_000, 3_965_001)));
+                "4 elements", p.print(Series.ofInt(1, -1, -10_000, 3_965_001)));
     }
 
     @Test
-    public void toString_Boolean() {
+    public void print_Boolean() {
         TabularPrinter p = new TabularPrinter(5, 10);
 
         assertEquals(System.lineSeparator() +
                 " true" + System.lineSeparator() +
                 "false" + System.lineSeparator() +
-                "2 elements", p.toString(Series.ofBool(true, false)));
+                "2 elements", p.print(Series.ofBool(true, false)));
     }
 
     @Test
-    public void toString_Full() {
+    public void print_Full() {
         TabularPrinter p = new TabularPrinter(5, 10);
 
         assertEquals(System.lineSeparator() +
@@ -50,11 +50,11 @@ public class TabularPrinter_SeriesTest {
                 "two  " + System.lineSeparator() +
                 "three" + System.lineSeparator() +
                 "four " + System.lineSeparator() +
-                "4 elements", p.toString(Series.of("one", "two", "three", "four")));
+                "4 elements", p.print(Series.of("one", "two", "three", "four")));
     }
 
     @Test
-    public void toString_TruncateRows() {
+    public void print_TruncateRows() {
 
         TabularPrinter p = new TabularPrinter(2, 10);
 
@@ -62,11 +62,11 @@ public class TabularPrinter_SeriesTest {
                 "one " + System.lineSeparator() +
                 "... " + System.lineSeparator() +
                 "four" + System.lineSeparator() +
-                "4 elements", p.toString(Series.of("one", "two", "three", "four")));
+                "4 elements", p.print(Series.of("one", "two", "three", "four")));
     }
 
     @Test
-    public void toString_TruncateColumns() {
+    public void print_TruncateColumns() {
         TabularPrinter p = new TabularPrinter(5, 4);
 
         assertEquals(System.lineSeparator() +
@@ -74,17 +74,17 @@ public class TabularPrinter_SeriesTest {
                 "two " + System.lineSeparator() +
                 "t..e" + System.lineSeparator() +
                 "four" + System.lineSeparator() +
-                "4 elements", p.toString(Series.of("one", "two", "three", "four")));
+                "4 elements", p.print(Series.of("one", "two", "three", "four")));
     }
 
     @Test
-    public void toString_EmptyStringsOnly() {
+    public void print_EmptyStringsOnly() {
 
         TabularPrinter p = new TabularPrinter(5, 10);
 
         assertEquals(System.lineSeparator() +
                 " " + System.lineSeparator() +
                 " " + System.lineSeparator() +
-                "2 elements", p.toString(Series.of("", "")));
+                "2 elements", p.print(Series.of("", "")));
     }
 }