+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #462 "Exp.shift(..)" in specialized expressions must return the type of that expression
 * #463 "Exp.as(..)" in specialized expressions must return the type of that expression
+* #464 Printer.toString(..) -> Printer.print(..)
 
 ## Release 2.0.0-M2